Pills do not show selected value(s) if disabled #12388 (#12555)
## Describe your changes
Before the change, when disabling a st.pills component it turned into a
faded gray color. The goal with the issue was to make still make it
faded but keep the same colors as when not disabled.

In styled-components.ts for BaseButton there is a style called
StyledButtonGroupBaseButton where the colors was set when a button is
disabled. Instead of changing color I when disabling a button I just
used opacity to make the faded look. Since StyledButtonGroupBaseButton
is used in both st.pills and st.segmented_control this change will
affect both of these components.

### disabled=False
This is how it looks when the button is not disabled (just like before)


https://github.com/user-attachments/assets/75bb71e9-2493-4c3b-bb1f-6f61ebc2c7ff

### disabled=True 
Before change:


https://github.com/user-attachments/assets/5f5e7d32-4d1a-4b8a-869a-d951e1fec00f

After change:


https://github.com/user-attachments/assets/9d039b62-d5ea-4560-8547-127f7dca25b6



<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link 
Closes https://github.com/streamlit/streamlit/issues/12388

## Testing Plan

I have made changes in e2e_playwright/st_pills_test.py and
e2e_playwright/st_segmented_control_test.py.
I updated the disabled tests to expect opacity instead of negating an
rgb(...) color, matching the new design.

- Unit Tests (JS and Python) ✅ 
- E2E Tests: st_segmented_control_test.py ✅ 
- E2E Tests: st_pills_test.py ✅ 


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Bob Nisco <bob.nisco@snowflake.com>