Makes fuzzy search case insensitive (#12849)
## Describe your changes

When working on #11309, I was working to provide more predictable case
sensitive search. It works in some cases (where the case sensitive
should come before a case insensitive), but it made fuzzy searching
across other instances worse (See #11724) In the end, we will go for
case insensitive search, because that's a little more intuitive as users
type and will be less great from a case sensitivity.

I think the scoring could be better in that case insensitive can be
slightly behind a case sensitive match, but that's a much tougher
problem to solve. This will alleviate user's concerns

## GitHub Issue Link (if applicable)

closes #11724

## Testing Plan

- Added a unit test based on the issue's use case

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.