Add Visitor pattern to AppNodes with a Debug Visitor (#12810)
## Describe your changes

This change introduces the [Visitor Design
Pattern](https://refactoring.guru/design-patterns/visitor) to our App
Node hierarchy. This allows having one file represent an operation
across all of the node types rather having to edit every node file (and
lose context).

To keep the PR small and not touch other methods, I created a
DebugVisitor that is designed to print out the node hierarchy (which I
think will be useful in debugging). It also demonstrates the purpose of
the design pattern.

## Testing Plan

- Unit Tests (JS and/or Python) - Adds JS unit tests for the Debug
Visitor and the AppRoot.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.