[fix] flaky content width container snapshot (#12906)
## Describe your changes

This test has been flaky due to a 1px fluctuation that is likely caused
by some rounding differences when the content width container is
combined with charts. The snapshot sizes must be consistent so we use a
fixed width wrapper container to keep the size consistent.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

Test only.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.