[chore] unskip test_layouts_container_various_elements in chromium (#12928)
## Describe your changes

This test was skipped in chromium on Friday, but since then a fix for
the flaky snapshot subtest has been merged
https://github.com/streamlit/streamlit/pull/12906

The other flaky snapshot subtest that we see failing is failing in
Webkit not Chromium so this skipping has no affect and we can unskip
this independently of resolving that issue.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

Test only. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.