Add optional uvloop support for better event loop performance (#13047)
## Describe your changes

Adds optional support for using uvloop as event loop. uvloop is an
ultra-fast, drop-in replacement of the built-in asyncio event loop. This
feature will get activated by having uvloop installed.

This change is a preparation for #12772 

## Testing Plan

- Added unit test.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <103002573+sfc-gh-lmasuch@users.noreply.github.com>
Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>