[fix] Ensure key is reflective of data arg in CCv2 instances (#12950)
## Describe your changes

- This fixes a few issues with how CCv2 handles the `data` arg,
especially as it relates to creating a `key`.
- Specifically, we weren't passing in the `data` arg to
`compute_and_register_element_id` before this change. Now we do so, but
in order to support the many different data shapes, we calculate a
stable identifier for the content first before passing it in.
- As a side-effect of ensuring the `data` param is stable, we move from
a UUID-based placeholder to an MD5-based placeholder for serialized
arrow data to ensure it is deterministic. Previously, it would be a new
UUID on each run, which is inefficient.
- Writes a bunch of unit tests that exercise the various `data` shapes.

## GitHub Issue Link (if applicable)

## Testing Plan

- Added unit tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.