Lower flaky-verify workflow to 25 runners (#13160)
## Describe your changes

The `flaky-verify` workflow is very useful but also somewhat expensive.
It stacked up 17k minutes (of the 8 core job) this month. Lowering this
to 25 runners, which verifies that the probability of workflow failure
is < 0.16% which should be more than fine for us:

<img width="737" height="239" alt="image"
src="https://github.com/user-attachments/assets/def47c68-c6de-444b-8b58-127681d7bc33"
/>

This change also adds another step to auto-remove the `flaky-verify`
label when run.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.