Refactor dataframe state handling into a dedicated hook (#13179)
## Describe your changes

Refactors the widget state (editing & selection) logic of the dataframe
component into a dedicated useWidgetState hook. This cleans-up some
complexity in Dataframe.tsx amd allows adding unit test for the hook.

This PR is a pure refactoring without any expected impact on the
behaviour.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.