Add layer of abstraction to generalize the CacheStat type (#13265)
Currently, the `_stcore/metrics` endpoint only returns data related to caches, including `st.session_state`,
`@st.cache_data`, and `@st.cache_resource`. In the SiS vNext world, there are various use cases that where
it would be useful to be able to expose more metrics from this endpoint.

We'll be both adding additional metrics and making some changes to the endpoint itself in subsequent
PRs, but as a first step, we'll just add another layer of abstraction to generalize the existing `CacheStatsProvider`
class.

In this PR, we
  * add a new `Stat` protocol with required properties that give us enough
     information to describe an OpenMetrics metric family
  * have the existing `CacheStat` class implement the new protocol
  * change the name of `CacheStatsProvider` to `StatsProvider`, and edit
     the signatures of `StatsManager` methods to now return a dictionary mapping from metric family name to
     lists of metrics rather than only returning lists of `CacheStat`s.
  * make corresponding downstream changes throughout the codebase so that
     things play nicely with our new `StatsManager.register_provider()` and `StatsManager.get_stats()`
     signatures.

We make one relatively small behavioral change to not include metrics comments if we don't have any
metrics to report, but aside from this, we don't make any behavioral changes in this PR.