Add on_release to st.cache_resource. (#13439)
## Describe your changes

Replace the TTLCache in ResourceCache with a TTLCleanupCache.

Add on_release to the st.cache_resource API, and plumb it through to the
TTLCleanupCache.

Update st.cache_resource.clear_all to clear the cache directly instead
of just GCing it to ensure release functions are called.

Implements feature request in #8764 .

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/8674

## Testing Plan

- Unit Tests (JS and/or Python)

Implemented.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>